refactor: chart configuration location
